remove owner and repo from inputs
